requirements.txt
asgiref==3.3.4
click==8.0.1
fastapi==0.65.2
h11==0.12.0
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pydantic==1.8.2
pytesseract==0.3.7
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
